april15/normal/damjan-cvetko/challenge2/secure_test.go
package main

import (
	"fmt"
	"io"
	"testing"
)

func TestReadWriterPartialPing(t *testing.T) {
	priv, pub := &[32]byte{'p', 'r', 'i', 'v'}, &[32]byte{'p', 'u', 'b'}

	r, w := io.Pipe()
	secureR := NewSecureReader(r, priv, pub)

	secureW := NewSecureWriter(w, priv, pub)

	// Encrypt hello world
	go func() {
		fmt.Fprintf(secureW, "hello world\n")
		w.Close()
	}()

	// Decrypt message
	buf := make([]byte, 1024)
	n, err := secureR.Read(buf[:3])
	if err != nil {
		t.Fatal(err)
	}
	m, err := secureR.Read(buf[3:])
	if err != nil {
		t.Fatal(err)
	}
	buf = buf[:n+m]

	// Make sure we have hello world back
	if res := string(buf); res != "hello world\n" {
		t.Fatalf("Unexpected result: %s != %s", res, "hello world")
	}
}
